part 10

+++ domain/src/main/java/com/codewithfk/domain/usecase/CartSummaryUseCase.kt
@@ -3,5 +3,5 @@ package com.codewithfk.domain.usecase
 import com.codewithfk.domain.repository.CartRepository
 
 class CartSummaryUseCase (private val repository: CartRepository) {
-    suspend fun execute(userId: Int) = repository.getCartSummary(userId)
+    suspend fun execute(userId: Long) = repository.getCartSummary(userId)
 }
